requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf==6.6.0
python-docx==1.2.0
